fix binding tests, get contract pinning hashes generated for mainnet and testnet contract dependencies
